refactor: move the concern out of app/modules and de-flake its spec
CONTRIBUTING forbids new files in app/modules/. Also replace the spec's sleep-based thread
ordering with a latch the stub signals once the lock is held, so a scheduler stall can no
longer redden a correct build, and clean up the paper_trail rows the non-transactional
example leaves behind.
